going back to less and combine other way
