+++ models/__init__.py
@@ -1,23 +1,12 @@
 #!/usr/bin/python3
-'''Package initializer'''
-
-from models.base_model import BaseModel
-from models.user import User
-from models.review import Review
-from models.place import Place
-from models.state import State
-from models.city import City
-from models.amenity import Amenity
-from models.engine.file_storage import FileStorage
+"""This module instantiates an object of class FileStorage"""
 import os
+from models.engine import file_storage
+from models.engine import db_storage
 
 
-classes = {"User": User, "BaseModel": BaseModel, "Place": Place,
-           "State": State, "City": City, "Amenity": Amenity, "Review": Review}
-
-if os.getenv('HBNB_TYPE_STORAGE') == 'db':
-    from models.engine.db_storage import DBStorage
-    storage = DBStorage()
+if os.getenv("HBNB_TYPE_STORAGE") == "db":
+    storage = db_storage.DBStorage()
 else:
-    storage = FileStorage()
+    storage = file_storage.FileStorage()
 storage.reload()
