+++ models/base_model.py
@@ -1,67 +1,69 @@
 #!/usr/bin/python3
-"""BaseModel class"""
+"""This module defines a base class for all models in our hbnb clone"""
 import uuid
 from datetime import datetime
-import models
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import Column, Integer, String, DateTime
-import os
-
+from sqlalchemy import Column, String, DateTime
 
 Base = declarative_base()
 
 
 class BaseModel:
-    """Base class for other classes used."""
+    """A base class for all hbnb models"""
+
     id = Column(String(60), nullable=False, primary_key=True)
     created_at = Column(DateTime, nullable=False, default=datetime.utcnow())
     updated_at = Column(DateTime, nullable=False, default=datetime.utcnow())
 
     def __init__(self, *args, **kwargs):
-        """__init__ method for BaseModel class"""
-        if kwargs:
-            for name, value in kwargs.items():
-                if name != '__class__':
-                    if name == 'created_at' or name == 'updated_at':
-                        value = datetime.strptime(
-                            value, "%Y-%m-%dT%H:%M:%S.%f")
-                    setattr(self, name, value)
-            if 'id' not in kwargs:
-                setattr(self, 'id', str(uuid.uuid4()))
-                setattr(self, 'created_at', datetime.utcnow())
-                self.updated_at = self.created_at
-        else:
+        """Instatntiates a new model"""
+        if not kwargs:
             self.id = str(uuid.uuid4())
-            self.created_at = datetime.utcnow()
-            self.updated_at = self.created_at
+            self.created_at = datetime.now()
+            self.updated_at = datetime.now()
+        else:
+            for key, value in kwargs.items():
+                if key not in ["__class__", "created_at", "updated_at"]:
+                    setattr(self, key, value)
+            if "__class__" in kwargs:
+                del kwargs['__class__']
+            if "update_at" in kwargs:
+                kwargs['updated_at'] = datetime.strptime(kwargs['updated_at'],
+                                                     '%Y-%m-%dT%H:%M:%S.%f')
+            else:
+                self.updated_at = datetime.now()
+            if "created_at" in kwargs:
+                kwargs['created_at'] = datetime.strptime(kwargs['created_at'],
+                                                     '%Y-%m-%dT%H:%M:%S.%f')
+            else:
+                self.created_at = datetime.now()
+            self.__dict__.update(kwargs)
 
     def __str__(self):
-        """Return string of BaseModel class"""
-        base_dict = dict(self.__dict__)
-        base_dict.pop('_sa_instance_state', None)
-        return ("[{}] ({}) {}".format(self.__class__.__name__,
-                                      self.id, base_dict))
-
-    def __repr__(self):
-        """Return string of BaseModel class"""
-        return ("[{}] ({}) {}".format(self.__class__.__name__,
-                                      self.id, self.__dict__))
+        """Returns a string representation of the instance"""
+        cls = (str(type(self)).split('.')[-1]).split('\'')[0]
+        return '[{}] ({}) {}'.format(cls, self.id, self.__dict__)
 
     def save(self):
-        """save with new."""
+        """Updates updated_at with current time when instance is changed"""
+        from models import storage
         self.updated_at = datetime.now()
-        models.storage.new(self)
-        models.storage.save()
+        if self not in storage._DBStorage__session:
+            storage.new(self)
+        storage.save()
 
     def to_dict(self):
-        """Return dict of BaseModel class. """
-        cp_dct = dict(self.__dict__)
-        cp_dct['__class__'] = type(self).__name__
-        cp_dct['updated_at'] = cp_dct['updated_at'].isoformat()
-        cp_dct['created_at'] = cp_dct['created_at'].isoformat()
-        cp_dct.pop('_sa_instance_state', None)
-        return (cp_dct)
+        """Convert instance into dict format"""
+        dictionary = {}
+        dictionary.update(self.__dict__)
+        dictionary.update({'__class__':
+                          (str(type(self)).split('.')[-1]).split('\'')[0]})
+        dictionary['created_at'] = self.created_at.isoformat()
+        dictionary['updated_at'] = self.updated_at.isoformat()
+        dictionary.pop("_sa_instance_state", None)
+        return dictionary
 
     def delete(self):
-        """deletes the current storage"""
-        models.storage.delete(self)
+        """"""
+        from models import storage
+        storage.delete(self)
