+++ models/engine/file_storage.py
@@ -1,63 +1,63 @@
 #!/usr/bin/python3
-"""Define class FileStorage"""
+"""This module defines a class to manage file storage for hbnb clone"""
 import json
-import models
+from models.base_model import BaseModel
+from models.user import User
+from models.place import Place
+from models.state import State
+from models.city import City
+from models.amenity import Amenity
+from models.review import Review
 
 
 class FileStorage:
-    """Serializes instances to JSON file and deserializes to JSON file"""
-    __file_path = "file.json"
+    """This class manages storage of hbnb models in JSON format"""
+    __file_path = 'file.json'
     __objects = {}
 
     def all(self, cls=None):
-        """Return the dictionary"""
-        obje = {}
-        if cls is None:
-            return (self.__objects)
-        else:
+        """Returns a dictionary of models currently in storage of class"""
+        if cls is not None:
             if type(cls) is str:
-                cls = models.classes[cls]
-            for key, val in self.__objects.items():
-                if cls.__name__ == val.__class__.__name__:
-                    obje[key] = val
-            return (obje)
+                cls = eval(cls)
+            cls_dict = {}
+            for key, value in self.__objects.items():
+                if type(value) is cls:
+                    cls_dict[key] = value
+            return cls_dict
+        return self.__objects
 
     def new(self, obj):
-        """Set objects with key """
-        key = str(obj.__class__.__name__) + "." + str(obj.id)
-        value_dict = obj
-        FileStorage.__objects[key] = value_dict
+        """Adds new object to storage dictionary"""
+        self.__objects["{}.{}".format(type(obj).__name__, obj.id)] = obj
 
     def save(self):
-        """Serializes to JSON file. """
-        objects_dict = {}
-        for key, val in FileStorage.__objects.items():
-            objects_dict[key] = val.to_dict()
-
-        with open(FileStorage.__file_path, mode='w', encoding="UTF8") as fd:
-            json.dump(objects_dict, fd)
-
-    def delete(self, obj=None):
-        """delete objects"""
-        if not obj:
-            return
-        key = '{}.{}'.format(type(obj).__name__, obj.id)
-        if key in self.__objects:
-            del self.__objects[key]
-            self.save()
+        """Saves storage dictionary to file"""
+        t_dict = {obj: self.__objects[obj].to_dict()
+                  for obj in self.__objects.keys()}
+        with open(FileStorage.__file_path, 'w') as file:
+            json.dump(t_dict, file)
 
     def reload(self):
-        """Deserializes JSON file to an object."""
+        """Loads dictionary for storage from file"""
+
+        classes = {
+                    'BaseModel': BaseModel, 'User': User, 'Place': Place,
+                    'State': State, 'City': City, 'Amenity': Amenity,
+                    'Review': Review
+                  }
         try:
-            with open(FileStorage.__file_path, encoding="UTF8") as fd:
-                FileStorage.__objects = json.load(fd)
-            for key, val in FileStorage.__objects.items():
-                class_name = val["__class__"]
-                class_name = models.classes[class_name]
-                FileStorage.__objects[key] = class_name(**val)
+            temp = {}
+            with open(FileStorage.__file_path, 'r') as f:
+                temp = json.load(f)
+                for key, val in temp.items():
+                    self.all()[key] = classes[val['__class__']](**val)
         except FileNotFoundError:
             pass
 
-    def close(self):
-        """Reload object from JSON file"""
-        self.reload()
+    def delete(self, obj=None):
+        """ Deletes a object from __objects if it exists """
+        try:
+            del self.__objects["{}.{}".format(type(obj).__name__, obj.id)]
+        except (AttributeError, KeyError):
+            pass
