--- models_lee/engine/file_storage.py
@@ -1,63 +1,0 @@
-#!/usr/bin/python3
-"""This module defines a class to manage file storage for hbnb clone"""
-import json
-from models.base_model import BaseModel
-from models.user import User
-from models.place import Place
-from models.state import State
-from models.city import City
-from models.amenity import Amenity
-from models.review import Review
-
-
-class FileStorage:
-    """This class manages storage of hbnb models in JSON format"""
-    __file_path = 'file.json'
-    __objects = {}
-
-    def all(self, cls=None):
-        """Returns a dictionary of models currently in storage of class"""
-        if cls is not None:
-            if type(cls) is str:
-                cls = eval(cls)
-            cls_dict = {}
-            for key, value in self.__objects.items():
-                if type(value) is cls:
-                    cls_dict[key] = value
-            return cls_dict
-        return self.__objects
-
-    def new(self, obj):
-        """Adds new object to storage dictionary"""
-        self.__objects["{}.{}".format(type(obj).__name__, obj.id)] = obj
-
-    def save(self):
-        """Saves storage dictionary to file"""
-        t_dict = {obj: self.__objects[obj].to_dict()
-                  for obj in self.__objects.keys()}
-        with open(FileStorage.__file_path, 'w') as file:
-            json.dump(t_dict, file)
-
-    def reload(self):
-        """Loads dictionary for storage from file"""
-
-        classes = {
-                    'BaseModel': BaseModel, 'User': User, 'Place': Place,
-                    'State': State, 'City': City, 'Amenity': Amenity,
-                    'Review': Review
-                  }
-        try:
-            temp = {}
-            with open(FileStorage.__file_path, 'r') as f:
-                temp = json.load(f)
-                for key, val in temp.items():
-                    self.all()[key] = classes[val['__class__']](**val)
-        except FileNotFoundError:
-            pass
-
-    def delete(self, obj=None):
-        """ Deletes a object from __objects if it exists """
-        try:
-            del self.__objects["{}.{}".format(type(obj).__name__, obj.id)]
-        except (AttributeError, KeyError):
-            pass
