+++ models_matt/city.py
@@ -1,0 +1,9 @@
+#!/usr/bin/python3
+""" City Module for HBNB project """
+from models.base_model import BaseModel
+
+
+class City(BaseModel):
+    """ The city class, contains state ID and name """
+    state_id = ""
+    name = ""
