+++ models_matt/state.py
@@ -1,0 +1,8 @@
+#!/usr/bin/python3
+""" State Module for HBNB project """
+from models.base_model import BaseModel
+
+
+class State(BaseModel):
+    """ State class """
+    name = ""
